change integer type to long in id and add PaymentRepository and SendPaymentRequest

+++ src/main/java/com/StefanKuchta/BankApp/db/mapper/AccountRowMapper.java
@@ -10,8 +10,8 @@ public class AccountRowMapper implements RowMapper<Account> {
 
     public Account mapRow(ResultSet rs, int rowNum) throws SQLException {
         Account account = new Account();
-        account.setId(rs.getInt("id"));
-        account.setUserId(rs.getInt("user_id"));
+        account.setId(rs.getLong("id"));
+        account.setUserId(rs.getLong("user_id"));
         account.setIban(rs.getString("iban"));
         account.setName(rs.getString("name"));
         account.setBalance(rs.getDouble("balance"));
