+++ src/main/java/com/StefanKuchta/BankApp/db/repository/AccountRepository.java
@@ -2,7 +2,7 @@ package com.StefanKuchta.BankApp.db.repository;
 
 import com.StefanKuchta.BankApp.db.mapper.AccountRowMapper;
 import com.StefanKuchta.BankApp.domain.Account;
-import com.StefanKuchta.BankApp.service.functions.IbanGenerator;
+import com.StefanKuchta.BankApp.db.service.functions.IbanGenerator;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.PreparedStatementCreator;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
@@ -31,12 +31,12 @@ public class AccountRepository {
         return jdbcTemplate.query(sql, accountRowMapper);
     }
 
-    public Account getAccountById(int id) {
+    public Account getAccountById(long id) {
         final String sql = "SELECT * FROM account WHERE account.id = " + id;
         return jdbcTemplate.queryForObject(sql, accountRowMapper);
     }
 
-    public Integer createAccount(Account account) {
+    public Long createAccount(Account account) {
         final String sql = "INSERT INTO account (user_id, iban, name, balance) VALUES (?,?,?,?)";
         KeyHolder keyHolder = new GeneratedKeyHolder();
 
@@ -44,7 +44,7 @@ public class AccountRepository {
             @Override
             public PreparedStatement createPreparedStatement(Connection con) throws SQLException {
                 PreparedStatement ps = con.prepareStatement(sql, PreparedStatement.RETURN_GENERATED_KEYS);
-                ps.setInt(1, account.getUserId());
+                ps.setLong(1, account.getUserId());
                 ps.setString(2, ibanGenerator.generateIban());
                 ps.setString(3, account.getName());
                 ps.setDouble(4, 0);
@@ -52,7 +52,7 @@ public class AccountRepository {
             }
         }, keyHolder);
         if (keyHolder.getKey() != null) {
-            return keyHolder.getKey().intValue();
+            return keyHolder.getKey().longValue();
         } else {
             return null;
         }
